+++ helpers/time_benchmark_test.go
@@ -45,3 +45,9 @@ func BenchmarkParseHour(b *testing.B) {
 		_, _ = ParseHour("12:03")
 	}
 }
+
+func BenchmarkParseHourMinute(b *testing.B) {
+	for b.Loop() {
+		_, _, _ = ParseHourMinute("12:03")
+	}
+}